+++ src/Schema/Directives/HasOneDirective.php
@@ -23,13 +23,13 @@ class HasOneDirective extends RelationDirective implements FieldResolver, Define
 """
 Corresponds to [the Eloquent relationship HasOne](https://laravel.com/docs/eloquent-relationships#one-to-one).
 """
-directive @hasOne(      
+directive @hasOne(
   """
   Specify the relationship method name in the model class,
   if it is named different from the field in the schema.
   """
   relation: String
-  
+
   """
   Apply scopes to the underlying query.
   """