+++ src/Schema/Directives/InterfaceDirective.php
@@ -22,7 +22,7 @@ class InterfaceDirective extends BaseDirective implements DefinedDirective
 """
 Use a custom resolver to determine the concrete type of an interface.
 """
-directive @interface(      
+directive @interface(
   """
   Reference to a custom type-resolver function.
   Consists of two parts: a class name and a method name, seperated by an `@` symbol.