+++ CHANGELOG.md
@@ -14,15 +14,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow existing mutation directives `@create`, `@update`, `@upsert` and `@delete` to function
   as nested arg resolvers https://github.com/nuwave/lighthouse/pull/899
 - Validate at schema build time that the `apply` argument `@rules` is an array https://github.com/nuwave/lighthouse/pull/1092
+- Add support in `@whereConstraints` for IN, IS NULL and BETWEEN operators https://github.com/nuwave/lighthouse/pull/1099
 - Add ability to define pivot data on nested mutations within `sync`, `syncWithoutDetach` and `connect` https://github.com/nuwave/lighthouse/pull/1110
 
 ### Changed
 
 - Remove `\Nuwave\Lighthouse\Execution\MutationExecutor` in favour of modular nested arg resolvers https://github.com/nuwave/lighthouse/pull/899
+- Register the operator enum for `@whereConstraints` programmatically and allow
+  overwriting it through a service provider https://github.com/nuwave/lighthouse/pull/1099
 
 ### Fixed
 
 - Enable chained rule provider directives (`ProvidesRules`) to merge the rules before validating https://github.com/nuwave/lighthouse/pull/1082
+- Apply nested `OR` conditions in `@whereConstraints` correctly https://github.com/nuwave/lighthouse/pull/1099
 
 ## [4.7.2](https://github.com/nuwave/lighthouse/compare/v4.7.1...v4.7.2)